High resolution sine wave with possibility to reduce resolution for the final timer. Sine table stored in 16bit integer. VREF_MAX value imported based on SINE_MAX and VREF_SCALER.
Ability to reduce counter maximum results in faster PWM and smoother low pass filtering of the Voltage reference which reduced hissing sound.

Also, the A4950 nonlinearity was accounted for in sineTable, but the difference in values is only up to 2% and there is naturally no noticeable smoothness improvement. More smoothness could be achieved by introducing torque ripple compensation due to motor saliency.